Bugs in quick filter and small refactoring (#20)
