+++ src/components/accounts/constants.js
@@ -1,0 +1,12 @@
+// eslint-disable-next-line
+export const TABLE_HEADERS = [
+  'type',
+  'name',
+  'balance',
+  'min-balance',
+  'max-balance',
+  'destination-rate-limit',
+  'origination-capacity',
+  'termination-capacity',
+  'total-capacity',
+];
