+++ src/components/cdrs/Cdrs.vue
@@ -14,14 +14,12 @@
       </template>
       <template v-slot:quickFilter>
         <div class="quickfilter">
-          <div>
-            <b-link>Apply custom filters</b-link>, or
-            <QuickTableFilter
-              :get-data="getCdrs"
-              :on-reset="onReset"
-              :date-range="dateRange"
-            />
-          </div>
+          <b-link>Apply custom filters</b-link>, or
+          <QuickTableFilter
+            :get-data="getCdrs"
+            :on-reset="onReset"
+            :date-range="dateRange"
+          />
         </div>
       </template>
     </DataTable>
@@ -29,11 +27,13 @@
 </template>
 
 <script>
-import { isEmpty, flow } from 'lodash';
+import { isEmpty, flow, get } from 'lodash';
 
 import utils from '../../utils';
 import DataTable from '../DataTable/DataTable';
 import QuickTableFilter from '../QuickTableFilter/QuickTableFilter';
+import { CDRS } from '../../constants';
+import { TABLE_HEADERS } from './constants';
 
 import 'vue2-daterange-picker/dist/vue2-daterange-picker.css';
 
@@ -49,145 +49,7 @@ export default {
     return {
       dateRange: utils.getLast24Hours(),
       // Table fields
-      fields: [
-        {
-          key: 'time-start',
-          label: 'Start Time',
-          sortable: true,
-        },
-        {
-          key: 'time-connect',
-          label: 'Connect Time',
-        },
-        {
-          key: 'time-end',
-          label: 'End Time',
-        },
-        {
-          key: 'duration',
-          label: 'Duration',
-        },
-        {
-          key: 'success',
-          label: 'Success',
-        },
-        {
-          key: 'destination-initial-interval',
-          label: 'Destination Initial Interval',
-        },
-        {
-          key: 'destination-initial-rate',
-          label: 'Destination Initial Rate',
-        },
-        {
-          key: 'destination-next-interval',
-          label: 'Destination Next Initial',
-        },
-        {
-          key: 'destination-next-rate',
-          label: 'Destination Next Rate',
-        },
-        {
-          key: 'destination-fee',
-          label: 'Destination Fee',
-        },
-        {
-          key: 'customer-price',
-          label: 'Customer Price',
-        },
-        {
-          key: 'src-name-in',
-          label: 'Src Name In',
-        },
-        {
-          key: 'src-prefix-in',
-          label: 'Src Prefix In',
-        },
-        {
-          key: 'from-domain',
-          label: 'From Domain',
-        },
-        {
-          key: 'dst-prefix-in',
-          label: 'Dst Prefix In',
-        },
-        {
-          key: 'to-domain',
-          label: 'To Domain',
-        },
-        {
-          key: 'ruri-domain',
-          label: 'R-URI Domain',
-        },
-        {
-          key: 'diversion-in',
-          label: 'Diversion In',
-        },
-        {
-          key: 'local-tag',
-          label: 'Local Tag',
-        },
-        {
-          key: 'lega-disconnect-code',
-          label: 'Leg A Disconnect Code',
-        },
-        {
-          key: 'lega-disconnect-reason',
-          label: 'Leg A Disconnect Reason',
-        },
-        {
-          key: 'lega-rx-payloads',
-          label: 'Leg A Rx Payloads',
-        },
-        {
-          key: 'lega-tx-payloads',
-          label: 'Leg A Tx Payloads',
-        },
-        {
-          key: 'auth-orig-transport-protocol-id',
-          label: 'Auth Origin Transport Protocol Id',
-        },
-        {
-          key: 'auth-orig-ip',
-          label: 'Auth Origin Ip',
-        },
-        {
-          key: 'auth-orig-port',
-          label: 'Auth Origin Port',
-        },
-        {
-          key: 'lega-rx-bytes',
-          label: 'Leg A Rx Bytes',
-        },
-        {
-          key: 'lega-tx-bytes',
-          label: 'Leg A Tx Bytes',
-        },
-        {
-          key: 'lega-rx-decode-errs',
-          label: 'Leg A Rx Decode Errors',
-        },
-        {
-          key: 'lega-rx-no-buf-errs',
-          label: 'Leg A Rx No Buf Errors',
-        },
-        {
-          key: 'lega-rx-parse-errs',
-          label: 'Leg A Rx Parse Errors',
-        },
-        {
-          key: 'src-prefix-routing',
-          label: 'Src Prefix Routing',
-        },
-        {
-          key: 'dst-prefix-routing',
-          label: 'Dst Prefix Routing',
-        },
-        {
-          key: 'destination-prefix',
-          label: 'Destination Prefix',
-        },
-      ],
+      fields: TABLE_HEADERS,
     };
   },
   computed: {
@@ -201,19 +63,18 @@ export default {
       };
     },
     rows() {
-      return this.$store.getters.cdrs && this.$store.getters.cdrs.meta
-        ? this.$store.getters.cdrs.meta['total-count'] : 0;
+      return get(this.$store.getters, ['cdrs', 'meta', 'total-count'], 0);
     },
   },
   created() {
     if (isEmpty(this.$store.getters.cdrFilter)) {
-      this.$store.dispatch('setCdrFilter', this.filterValue);
+      this.$store.dispatch(CDRS.ACTIONS.SET_CDRS_FILTER, this.filterValue);
     }
     this.getCdrs();
   },
   methods: {
     getCdrs(pageNumber) {
-      this.$store.dispatch('getCdrs', pageNumber);
+      this.$store.dispatch(CDRS.ACTIONS.GET_CDRS, pageNumber);
     },
     onReset() {
       this.$data.dateRange = utils.getLast24Hours();
