+++ src/components/rates/constants.js
@@ -1,0 +1,43 @@
+// eslint-disable-next-line
+export const TABLE_HEADERS = [
+  {
+    key: 'connect-fee',
+    label: 'Connect fee',
+  },
+  {
+    key: 'initial-interval',
+    label: 'Initial interval',
+  },
+  {
+    key: 'initial-rate',
+    label: 'Initial rate',
+  },
+  {
+    key: 'network-prefix',
+    label: 'Network prefix',
+  },
+  {
+    key: 'next-interval',
+    label: 'Next interval',
+  },
+  {
+    key: 'next-rate',
+    label: 'Next rate',
+  },
+  {
+    key: 'prefix',
+    label: 'Prefix',
+  },
+  {
+    key: 'reject-calls',
+    label: 'Reject calls',
+  },
+  {
+    key: 'valid-from',
+    label: 'Valid from',
+  },
+  {
+    key: 'valid-till',
+    label: 'Valid till',
+  },
+];
