+++ src/constants/common.js
@@ -1,0 +1,5 @@
+// eslint-disable-next-line
+export const NOTIFICATION_TYPES = {
+  SUCCESS: 'success',
+  ERROR: 'error',
+};
